requirements.txt
Django
djangorestframework
psycopg2-binary
transformers        # LLaMA
torch
llama_cpp_python
django-cors-headers
python-decouple
numpy
matplotlib
futures